feat : 1.환경변수가 local인 환경인 경우, aws ssm 안 거치도록 분리. 2. 1.에 의해 environment_loader.py 추출 3. environment 패키지 따로 만들고, llm_interface.py, aws_interface.py,environment_loader.py 이동.
