+++ moview/modules/answer_evaluator/interview_answer_scorer.py
@@ -5,7 +5,7 @@ from langchain.chains import ConversationChain
 from langchain.chains.router.llm_router import LLMRouterChain, RouterOutputParser
 from langchain.chains.router.multi_prompt_prompt import MULTI_PROMPT_ROUTER_TEMPLATE
 from moview.modules.prompt_loader.prompt_loader import PromptLoader
-from moview.utils.llm_interface import LLMModelFactory
+from moview.envrionment.llm_interface import LLMModelFactory
 
 CATEGORIES = ["Behavioral Questions", "Situational Questions", "Technical Job-related Questions",
               "Cultural Fit Questions", "Personal Character Questions"]
