+++ moview/modules/initial_input/initial_question_giver.py
@@ -8,7 +8,7 @@ from langchain.prompts.chat import (
     HumanMessagePromptTemplate
 )
 from moview.modules.prompt_loader.prompt_loader import PromptLoader
-from moview.utils.llm_interface import LLMModelFactory
+from moview.envrionment.llm_interface import LLMModelFactory
 
 
 class InitialQuestionGiver:
